Bug 17173: Quick jump to subfield edition form
There are too many clicks to jump to the edit form of a specific
subfield.

This patch adds a hidden feature (i.e. will not be polished or
documented).

When searching for a tag field from a marc framework structure, it's not
possible to search for a tag subfield to jump directly to the subfield
edition form.

Test plan:
On /admin/marctagstructure.pl?frameworkcode=FRAMEWORKCODE
Search for a subfield (952$d)
you will be redirected to the subfield edition form

There is no check if the subfield does not exist.

Followed test plan, works as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>