Bug 17194 - When edit record, Button "Z39.50/SRU search" not work
Bug 15859 introduced a problem in the basic cataloging editor by
removing a form field which the Z39.50 search button tries to use when
triggering the search window pop-up. This patch corrects the error by
changing where the script looks for the required framework code.

To test, apply the patch and got to cataloging.

- Edit a record which uses the default framework.
  - Confirm that the Z39.50/SRU search button works correctly to trigger
    the pop-up window.
- Repeat the test with a record which uses a different framework.

Signed-off-by: Liz Rea <liz@catalyst.net.nz>
Working again, great!

Signed-off-by: Andreas Roussos <arouss1980@gmail.com>
Patch works fine.

Signed-off-by: Katrin Fischer  <katrin.fischer@bsz-bw.de>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>