Bug 7679: Follow-up to fix group by hour
Grouping by hour does not work with the collates.
This patch removes them after talking to Julian on IRC.

This partially reverts one of the earlier patches:
Bug 7679: Display all table headers (case and accent
 sensitive)

One could argue that this is more of a data problem,
than something we should fix/'hide' with code.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>