Bug 17537: Fix valid-templates.t for some include files
3 include files do not pass the template validation tests:
 - koha-tmpl/intranet-tmpl/prog/en/includes/admin-items-search-field-form.inc
 - koha-tmpl/intranet-tmpl/prog/en/includes/subscriptions-search.inc
 - koha-tmpl/opac-tmpl/bootstrap/en/includes/opac-topissues.inc

This is because they process a block which has not been declared before.
As they are include files, they cannot stand on their own.
We could have added them to the exclude file list of xt/author/valid-templates.t
but I think it's better to keep them into the validation loop.

Test plan:
  prove xt/author/valid-templates.t
should return green
And the library dropdown list should be correctly populated on the pages these files are included
(opac-topissues for instance)

Signed-off-by: Claire Gravely <claire_gravely@hotmail.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>