Bug 17545 [Revised] Make "Add biblio" not hidden by language chooser
This alternate patch changes the menu direction so that it "drops" up
instead of down. This avoids the problem of overlapping or being hidden
by the language chooser.

The disadvantage of this patch is that it is not global. However, I
think it's a better long-term solution to the problem.

To test, apply the patch and follow the issue description:

- Make sure more than one language is active for the staff client, so
  that the language chooser bar shows at the bottom of each page
- Do one or more Z39.50 searches, to make sure you have some records in
  the reservoir
- Go to "Cataloging" and do a "Catalog search", so that the reservoir is
  searched. Make sure your search is broad enough that some results from
  the reservoir is shown "below the fold"
- Scroll to the last result and click on the "Actions" menu

The direction of the menu should be up instead of down.

Signed-off-by: Magnus Enger <magnus@libriotech.no>
Popup menus on the cataloging search page now pop up instead
of down, which means they avoid being hidden by the language
chooser. Mission accimplished!

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>